more tests

+++ src/serialbuffer.cpp
@@ -36,7 +36,7 @@ SerialBuffer::putInt16(int16_t v)
     if (this->writeIdx + sizeof(v) - 1 >= this->sizeOfBuf)
 	return false;
 
-    uint16_t v2 = htons((uint8_t)v);
+    uint16_t v2 = htons((uint16_t)v);
     uint8_t b0 = v2 & 0xFF;
     uint8_t b1 = (v2 >> 8) & 0xFF;
     this->buf[ this->writeIdx++ ] = b0;
